moving lizmapRepository to namespace Projec
