+++ lizmap/modules/lizmap/classes/lizmapRepository.class.php
@@ -9,11 +9,12 @@
  *
  * @license Mozilla Public License : http://www.mozilla.org/MPL/
  */
+
+/**
+ * @deprecated
+ */
 class lizmapRepository
 {
-    // Lizmap configuration file path (relative to the path folder)
-    private $config = 'config/lizmapConfig.ini.php';
-
     /**
      * services properties.
      *
@@ -45,78 +46,24 @@ class lizmapRepository
         ),
     );
 
-    // Lizmap repository key
-    private $key = '';
-    // Lizmap repository configuration data
-    private $data = array();
     /**
-     * @var lizmapProject[] list of projects. keys are projects names
+     * @var \Lizmap\Project\Repository The repository instance
      */
-    protected $projectInstances = array();
-    // The configuration files folder path
-    private $varPath = '';
+    protected $repo;
 
-    protected $services;
-
-    protected $appContext;
-
-    /**
-     * lizmapRepository Constructor
-     * Do not call it, if you want to instanciate a lizmapRepository, you should
-     * do it with the lizmapServices::getLizmapRepository method.
-     *
-     * @param string         $key      the name of the repository
-     * @param array          $data     the repository data
-     * @param string         $varPath  the configuration files folder path
-     * @param lizmapServices $services
-     * @param Lizmap\App\AppContextInterface
-     * @param mixed $appContext
-     */
-    public function __construct($key, $data, $varPath, $services, $appContext)
+    public function __construct($key, $data, $varPath, $context, $services)
     {
-        $properties = self::getProperties();
-        $this->varPath = $varPath;
-        $this->services = $services;
-        $this->appContext = $appContext;
-
-        // Set each property
-        foreach ($properties as $property) {
-            if (array_key_exists($property, $data)) {
-                $this->data[$property] = $data[$property];
-            }
-        }
-        $this->key = $key;
+        $this->repo = new \Lizmap\Project\Repository($key, $data, $varPath, $context, $services);
     }
 
     public function getKey()
     {
-        return $this->key;
+        return $this->repo->getKey();
     }
 
     public function getPath()
     {
-        if ($this->data['path'] == '') {
-            return false;
-        }
-        // add a trailing slash if needed
-        if (!preg_match('#/$#', $this->data['path'])) {
-            $this->data['path'] .= '/';
-        }
-        $path = $this->data['path'];
-        // if path is relative, get full path
-        if ($this->data['path'][0] != '/' and $this->data['path'][1] != ':') {
-            $path = realpath($this->varPath.$this->data['path']).'/';
-        }
-        if (strstr($this->data['path'], './')) {
-            $path = realpath($this->data['path']).'/';
-        }
-        if (file_exists($path)) {
-            $this->data['path'] = $path;
-        } else {
-            return false;
-        }
-
-        return $this->data['path'];
+        return $this->repo->getPath();
     }
 
     public static function getProperties()
@@ -126,21 +73,17 @@ class lizmapRepository
 
     public function getRepoProperties()
     {
-        return self::$properties;
+        return self::$propertiesOptions;
     }
 
     public static function getPropertiesOptions()
     {
-        return self::$propertiesOptions;
+        return self::$repo->getPropertiesOptions();
     }
 
     public function getData($key)
     {
-        if (!array_key_exists($key, $this->data)) {
-            return null;
-        }
-
-        return $this->data[$key];
+        return $this->repo->getData($key);
     }
 
     /**
@@ -153,87 +96,16 @@ class lizmapRepository
      */
     public function update($data, $ini)
     {
-        // Set section
-        $section = 'repository:'.$this->key;
-
-        $modified = false;
-        // Modify the ini data for the repository
-        foreach ($data as $k => $v) {
-            if (in_array($k, self::getProperties())) {
-                // Set values in ini file
-                $ini->setValue($k, $v, $section);
-                // Modify lizmapConfigData
-                $this->data[$k] = $v;
-                $modified = true;
-            }
-        }
-
-        // Save the ini file
-        if ($modified) {
-            $ini->save();
-        }
-
-        return $modified;
+        return $this->repo->update($data, $ini);
     }
 
-    public function getProject($key)
+    public function getProject($key, $context, $services)
     {
-        if (isset($this->projectInstances[$key])) {
-            return $this->projectInstances[$key];
-        }
-
-        try {
-            $proj = new lizmapProject($key, $this, $this->appContext, $this->services);
-        } catch (UnknownLizmapProjectException $e) {
-            throw $e;
-        } catch (Exception $e) {
-            jLog::logEx($e, 'error');
-
-            return null;
-        }
-
-        $this->projectInstances[$key] = $proj;
-
-        return $proj;
+        return $this->repo->getProject($key, $context, $services);
     }
 
-    public function getProjects()
+    public function getProjects($context, $services)
     {
-        $projects = array();
-        $dir = $this->getPath();
-
-        if (is_dir($dir)) {
-            if ($dh = opendir($dir)) {
-                $cfgFiles = array();
-                $qgsFiles = array();
-                while (($file = readdir($dh)) !== false) {
-                    if (substr($file, -3) == 'cfg') {
-                        $cfgFiles[] = $file;
-                    }
-                    if (substr($file, -3) == 'qgs') {
-                        $qgsFiles[] = $file;
-                    }
-                }
-                closedir($dh);
-
-                foreach ($qgsFiles as $qgsFile) {
-                    $proj = null;
-                    if (in_array($qgsFile.'.cfg', $cfgFiles)) {
-                        try {
-                            $proj = $this->getProject(substr($qgsFile, 0, -4));
-                            if ($proj != null) {
-                                $projects[] = $proj;
-                            }
-                        } catch (UnknownLizmapProjectException $e) {
-                            jLog::logEx($e, 'error');
-                        } catch (Exception $e) {
-                            jLog::logEx($e, 'error');
-                        }
-                    }
-                }
-            }
-        }
-
-        return $projects;
+        return $this->repo->getProjects($context, $services);
     }
 }
